requirements/core.txt
# ============================================
# DeepTutor Core Dependencies (CLI minimum)
# ============================================
# Install: pip install -r requirements/core.txt
# This is the minimal set for CLI-only usage (~80MB)

# --- Config & templating ---
python-dotenv>=1.0.0
PyYAML>=6.0
jinja2>=3.1.0

# --- LLM ---
openai>=1.30.0
tiktoken>=0.5.0

# --- HTTP clients ---
aiohttp>=3.9.4
httpx>=0.27.0
requests>=2.32.2

# --- Async helpers ---
nest_asyncio>=1.5.8
tenacity>=8.0.0

# --- Data models ---
pydantic>=2.0.0
aiosqlite>=0.19.0

# --- CLI framework ---
typer[all]>=0.9.0
rich>=13.0.0
